backend/requirements.txt
alembic==1.15.2
bcrypt==4.0.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
faiss-cpu==1.7.4
filelock==3.18.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
fsspec==2025.3.2
google-ai-generativelanguage==0.4.0
google-api-core==2.25.0rc1
google-auth==2.40.1
google-generativeai==0.3.1
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.62.3
gunicorn==21.2.0
huggingface-hub==0.31.2
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.20.1
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==1.24.3
packaging==25.0
pandas==2.0.3
pillow==11.2.1
pluggy==1.6.0
proto-plus==1.26.1
protobuf==4.25.7
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
pytest==7.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==2.2.2
sentencepiece==0.2.0
six==1.17.0
SQLAlchemy==2.0.41
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.13.3
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
